Added Codebase
ATM there is no IR support but the keypad is fully functional.
Code commenting will be added soon. this code is a little messy and i am working on improvements.
For NEW users you only really need to change the Settings.cpp file. This will allow you to add the code you are going to need